requirements.txt
Flask
Flask-Cors
Flask-Migrate
Flask-SQLAlchemy
PyJWT
Werkzeug
python-dotenv
GitPython